aider: Added required TTL values to the description of the `ttl_sec` variable in `tfmods/linode_domain/variables.tf` and updated the `README.md` accordingly.
# Aider chat conversation:

USER: How would I add the required values (determined with validations) in the terraform-docs generated information?

ASSISTANT: To add the required values determined by validations to the `terraform-docs` generated information, you would need to update the `variables.tf` and `README.md` files for the respective Terraform modules. Since `terraform-docs` generates documentation based on the content of the `variables.tf` and `outputs.tf` files, you would need to ensure that the descriptions in the `variables.tf` file include the required values.

Let's start with the `tfmods/linode_domain/variables.tf` file, where we will add the required TTL values to the description of the `ttl_sec` variable.

```diff
--- tfmods/linode_domain/variables.tf
+++ tfmods/linode_domain/variables.tf
@@ ... @@
 variable "domains" {
   description = "Domain data."

   type = map(object({
     domain      = string
     soa_email   = string

     description = optional(string)
     expire_sec  = optional(number)
     group       = optional(string)
     refresh_sec = optional(number)
     retry_sec   = optional(number)
     status      = optional(string)
     tags        = optional(list(string))
-    ttl_sec     = optional(number)
+    ttl_sec     = optional(number, description = "The Time To Live (TTL) for the domain, in seconds. Valid values are 0, 30, 120, 300, 3600, 7200, 14400, 28800, 57600, 86400, 172800, 345600, 604800, 1209600, 2419200.")
   }))
```

Next, we will update the `README.md` for the `tfmods/linode_domain` module to reflect the changes in the `variables.tf` file. Since `terraform-docs` uses the content of the `variables.tf` file to generate the documentation, running `terraform-docs` again after updating the `variables.tf` file will update the `README.md` accordingly.

However, if you want to manually update the `README.md` file, you can do so by adding the required values to the description of the `ttl_sec` input in the documentation section.

```diff
--- tfmods/linode_domain/README.md
+++ tfmods/linode_domain/README.md
@@ ... @@
 ## Inputs

 | Name | Description | Type | Default | Required |
 |------|-------------|------|---------|:--------:|
 | <a name="input_domains"></a> [domains](#input\_domains) | Domain data. | <pre>map(object({<br>    domain      = string<br>    soa_email   = string<br><br>    description = optional(string)<br>    expire_sec  = optional(number)<br>    group       = optional(string)<br>    refresh_sec = optional(number)<br>    retry_sec   = optional(number)<br>    status      = optional(string)<br>    tags        = optional(list(string))<br>-   ttl_sec     = optional(number)<br>+   ttl_sec     = optional(number, description = "The Time To Live (TTL) for the domain, in seconds. Valid values are 0, 30, 120, 300, 3600, 7200, 14400, 28800, 57600, 86400, 172800, 345600, 604800, 1209600, 2419200.")<br>  }))</pre> | n/a | yes |
```

Please note that the actual `terraform-docs` command needs to be run to regenerate the documentation based on the updated `variables.tf` file. The diffs provided above are for manual updates to the files. If you have `terraform-docs` installed and configured, you can run it to automatically update the `README.md` files.